AX: Isolated tree: Crash in URL retrieval https://bugs.webkit.org/show_bug.cgi?id=208942 <rdar://problem/60337588>
Reviewed by Chris Fleizach.

Source/WebCore:

Test: accessibility/url-test.html

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::url const): Check for m_renderer
before dereferencing and re-use the associated Node. Also fixes the last
type check and downcast that were using different types.

LayoutTests:

Moved the url-test.html test to the root directory so that it is
executed in both Mac and iOS.

* accessibility/url-test-expected.txt: Renamed from LayoutTests/accessibility/ios-simulator/url-test-expected.txt.
* accessibility/url-test.html: Renamed from LayoutTests/accessibility/ios-simulator/url-test.html.
* platform/ios-simulator/TestExpectations:
* platform/win/TestExpectations:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258346 268f45cc-cd09-0410-ab3c-d52691b4dbfc