[LFC][TFC] Fill in the gaps with anonymous table cell boxes https://bugs.webkit.org/show_bug.cgi?id=209114 <rdar://problem/60458806>
Reviewed by Antti Koivisto.

Source/WebCore:

17.5 Visual layout of table contents
A "missing cell" is a cell in the row/column grid that is not
occupied by an element or pseudo-element. Missing cells are rendered
as if an anonymous table-cell box occupied their position in the grid.
(https://www.w3.org/TR/CSS22/tables.html)

This helps to keep TableGrid an actual grid.

Test: fast/layoutformattingcontext/table-missing-cells-simple.html

* layout/layouttree/LayoutTreeBuilder.cpp:
(WebCore::Layout::TreeBuilder::buildTableStructure):

LayoutTests:

* fast/layoutformattingcontext/table-missing-cells-simple-expected.html: Added.
* fast/layoutformattingcontext/table-missing-cells-simple.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258473 268f45cc-cd09-0410-ab3c-d52691b4dbfc