[Multicolumn] RenderListItem::positionListMarker should not fail when the list marker is inside a spanner. https://bugs.webkit.org/show_bug.cgi?id=209262 <rdar://problem/58447665>
Reviewed by Simon Fraser.

Source/WebCore:

When the list marker is in a column spanner and as a result it gets moved under the column flow, the
normal "let's find the list item by walking up on the ancestor chain" does not work anymore.
We need to check if this list marker is inside a spanner and climb up on the ancestor chain by
using the spanner placeholder position (see RenderListMarker::parentBox).
This patch also moves the marker's overflow computation from the list item to the marker.

Test: fast/multicol/list-item-marker-inside-column-spanner.html

* rendering/RenderListItem.cpp:
(WebCore::RenderListItem::addOverflowFromChildren):
(WebCore::RenderListItem::positionListMarker): Deleted.
* rendering/RenderListMarker.cpp:
(WebCore::RenderListMarker::parentBox):
(WebCore::RenderListMarker::addOverflowFromListMarker):
(WebCore::RenderListMarker::layout):
* rendering/RenderListMarker.h:

LayoutTests:

* fast/multicol/list-item-marker-inside-column-spanner-expected.txt: Added.
* fast/multicol/list-item-marker-inside-column-spanner.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258680 268f45cc-cd09-0410-ab3c-d52691b4dbfc