[iOS+macOS Debug] ASSERTION FAILED: m_effects.isEmpty() in WebCore::KeyframeEffectStack::~KeyframeEffectStack() https://bugs.webkit.org/show_bug.cgi?id=208718 <rdar://problem/59956433>
Patch by Antoine Quint <graouts@apple.com> on 2020-03-24
Reviewed by Darin Adler.

Source/WebCore:

This assertion is only valid in the case where elements with relevant animations are still alive when the document is
torn down. However, it's fine to have an Element be destroyed with a relevant animation while its document is still alive.

* animation/KeyframeEffectStack.cpp:
(WebCore::KeyframeEffectStack::~KeyframeEffectStack):

LayoutTests:

* platform/ios-wk2/TestExpectations:
* platform/mac/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258913 268f45cc-cd09-0410-ab3c-d52691b4dbfc