Clear the entropy bits in the encodedStructureBits when deallocating a structureID. https://bugs.webkit.org/show_bug.cgi?id=209632 <rdar://problem/60943876>
Reviewed by Saam Barati.

We currently only use a 32-bit offset in the StructureIDTable's StructureOrOffset.
Though we will never store an offset value that is near 32-bit in size, let alone
64-bit, there's no reason why we can't just use all 64-bits for the offset.
Doing so will also have the benefit of zero'ing out the entropy bits in the old
encodedStructureBits.  This guarantees that there's no chance of coalition between
a "freed" structureID's entropy bits and the entropy bits in a dead cell due to
GC bugs.

* runtime/StructureIDTable.h:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259107 268f45cc-cd09-0410-ab3c-d52691b4dbfc