Sideways jiggles when scrolling the shelves on beta.music.apple.com https://bugs.webkit.org/show_bug.cgi?id=209696 <rdar://problem/55092050>
Reviewed by Anders Carlsson.
Source/WebCore:

If a scroll snapping animation was running, EventHandler::platformNotifyIfEndGesture() would
reset the latching state. This was added in r190423, but not longer seems necessary
according to manual testing, and the passing layout test.

platformNotifyIfEndGesture() would be called at the end of the fingers-down scroll but
before momentum, and resetting latching here would cause the momentum events to go to
a new target, triggering incorrect scrolls.

Test: tiled-drawing/scrolling/scroll-snap/scroll-snap-phase-change-relatching.html

* page/mac/EventHandlerMac.mm:
(WebCore::EventHandler::platformNotifyIfEndGesture):

LayoutTests:

Test that sends scroll and momentum events to a vertically-scrolling overflow with snap-points,
which checked that the document didn't scroll.

* tiled-drawing/scrolling/scroll-snap/scroll-snap-phase-change-relatching-expected.txt: Added.
* tiled-drawing/scrolling/scroll-snap/scroll-snap-phase-change-relatching.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259162 268f45cc-cd09-0410-ab3c-d52691b4dbfc