Datalist option's label not used https://bugs.webkit.org/show_bug.cgi?id=201768 <rdar://problem/55361186>
Reviewed by Darin Adler.

Source/WebCore:

Refactor DataListSuggestionInformation's suggestions to include label text as well as values, and then adjust
TextFieldInputType::suggestions() to match label text as well as values for ports that are capable of showing
label text in datalist suggestion UI.

Test: fast/forms/datalist/datalist-option-labels.html

* html/DataListSuggestionInformation.h:

Introduce DataListSuggestion, a wrapper around a value and label. Currently, the list of datalist suggestions
is only a `Vector<String>`; change it to be a `Vector<DataListSuggestion>` instead.

(WebCore::DataListSuggestion::encode const):
(WebCore::DataListSuggestion::decode):
(WebCore::DataListSuggestionInformation::encode const):
(WebCore::DataListSuggestionInformation::decode):

Move encoding and decoding for DataListSuggestionInformation out of WebCoreArgumentCoders and into WebCore.

* html/TextFieldInputType.cpp:
(WebCore::TextFieldInputType::listAttributeTargetChanged):
(WebCore::TextFieldInputType::suggestions):

When computing suggestions, match label text in addition to values on ports that display label text in the
chrome; for the time being, this is only the case for macOS, but will be extended to iOS as well in a future
patch. Note that we don't plumb label text if it is already the same as the value, to avoid duplicate strings
from showing up.

(WebCore::TextFieldInputType::didCloseSuggestions):
* html/TextFieldInputType.h:
* loader/EmptyClients.h:
* page/ChromeClient.h:

Add a chrome client hook to return whether or not the client shows label text in its datalist UI.

* platform/DataListSuggestionsClient.h:

Source/WebKit:

Add support on macOS for showing option labels in datalist suggestions.

* Shared/WebCoreArgumentCoders.cpp:
(IPC::ArgumentCoder<DataListSuggestionInformation>::encode): Deleted.
(IPC::ArgumentCoder<DataListSuggestionInformation>::decode): Deleted.
* Shared/WebCoreArgumentCoders.h:

Remove WebCoreArgumentCoders logic for encoding and decoding DataListSuggestionInformation. See
DataListSuggestionInformation.h in WebCore for more detail.

* UIProcess/gtk/WebDataListSuggestionsDropdownGtk.cpp:
(WebKit::WebDataListSuggestionsDropdownGtk::show):

Tweak GTK code to adjust for the change from `String` to `DataListSuggestion`.

* UIProcess/ios/WebDataListSuggestionsDropdownIOS.mm:
(-[WKDataListSuggestionsControl didSelectOptionAtIndex:]):
(-[WKDataListSuggestionsControl textSuggestions]):
(-[WKDataListSuggestionsControl suggestionAtIndex:]):

Adjust some iOS codepaths to use DataListSuggestion::value as the value string to display.

* UIProcess/mac/WebDataListSuggestionsDropdownMac.mm:

Tweak several UI constants. A suggestion cell may now be either 20 or 40pt tall, depending on whether it has
label text to show.

Currently, the maximum combined height of the table view cells is 120 (not including spacing between cells and
vertical padding around the top and bottom of the table view), since the maximum number of cells to show is 6
and each cell is 20pt tall. Maintain this constant by making the maximum cell height 120, which accomodates
either three labeled cells, or 6 unlabeled cells (i.e. to match shipping behavior).

(-[WKDataListSuggestionView initWithFrame:]):
(-[WKDataListSuggestionView layout]):

Maintain two text fields or value and (optionally) label text: `_valueField` and `_labelField`. The value field
fills the bounds of the cell in the case where there is no label text, but fills only the top half of the cell
in the case where there is label text. The label field takes the bottom half of the cell in this case.

Additionally, add a divider view that may appear at the very bottom of each cell. This divider view is present
when one or more suggestions in the datalist are labeled.

(-[WKDataListSuggestionView setValue:label:]):

Renamed from -setText:. Add a label string argument as well.

(-[WKDataListSuggestionView setShouldShowBottomDivider:]):

Add getters and setters for the -shouldShowBottomDivider property, which can be used to make the divider view
visible or hidden.

(-[WKDataListSuggestionView shouldShowBottomDivider]):
(-[WKDataListSuggestionView setBackgroundStyle:]):

Use -[NSColor secondaryLabelColor] for the label text field.

(shouldShowDividersBetweenCells):

Add a helper method to determine whether the table view should be showing clear dividers between each item.
We only do so if there are one or more labels to be shown.

(-[WKDataListSuggestionsController initWithInformation:inView:]):
(-[WKDataListSuggestionsController currentSelectedString]):
(-[WKDataListSuggestionsController updateWithInformation:]):
(-[WKDataListSuggestionsController moveSelectionByDirection:]):

Drive-by fix: scroll to reveal each selected row when using the arrow keys to navigate between items.

(-[WKDataListSuggestionsController dropdownRectForElementRect:]):
(-[WKDataListSuggestionsController tableView:heightOfRow:]):

Return either `dropdownRowHeightWithoutLabel` or `dropdownRowHeightWithLabel`, depending on whether there is
label text to be shown in that suggestion cell.

(-[WKDataListSuggestionsController tableView:viewForTableColumn:row:]):
(-[WKDataListSuggestionView setText:]): Deleted.
* WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::canShowDataListSuggestionLabels const):
* WebProcess/WebCoreSupport/WebChromeClient.h:

Source/WebKitLegacy/mac:

See WebCore/ChangeLog and WebKit/ChangeLog for more details.

* WebCoreSupport/WebChromeClient.h:

LayoutTests:

Add a test to verify that when datalist suggestions incorporate option labels into the displayed text
suggestions, they are matched against when changing the value of the input field, and can be selected. This is
expected to be the case on macOS, but not on other platforms (for the time being).

* fast/forms/datalist/datalist-option-labels-expected.txt: Added.
* fast/forms/datalist/datalist-option-labels.html: Added.
* platform/gtk/fast/forms/datalist/datalist-option-labels-expected.txt: Added.
* platform/ios/fast/forms/datalist/datalist-option-labels-expected.txt: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259330 268f45cc-cd09-0410-ab3c-d52691b4dbfc