[PlayStation] Use OBJECT libraries for WebCore and PAL https://bugs.webkit.org/show_bug.cgi?id=209835
Reviewed by Ross Kirsling.

.:

When compiling WebCore in debug mode ranlib was having issues with the size of
libWebCore.a. This is similar to what happened with Visual Studio and WinCairo.

Additionally hidden visibility defaults are added for the port.

* Source/cmake/OptionsPlayStation.cmake:

Source/WebCore:

After moving to object libraries MediaStrategy's constructor and destructor were
being reported as linker errors despite being set to default. Moving the definitions
to a source file stopped these issues.

* Sources.txt:
* WebCore.xcodeproj/project.pbxproj:
* platform/MediaStrategy.cpp: Copied from Source/WebCore/platform/MediaStrategy.h.
* platform/MediaStrategy.h:

Source/WebKit:

Use the object libraries when building WebKit.

* PlatformPlayStation.cmake:

Tools:

Use the object libraries when compiling TestWebCore.

* TestWebKitAPI/PlatformPlayStation.cmake:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259358 268f45cc-cd09-0410-ab3c-d52691b4dbfc