Video previews on shutterstock.com don't play when tapped on iPadOS https://bugs.webkit.org/show_bug.cgi?id=209903 <rdar://problem/58844166>
Reviewed by Wenson Hsieh.

When tapping on a video preview after searching for a video on shutterstock.com, for instance on https://www.shutterstock.com/video/search/people,
we correctly enter the hover state thanks to our content observation heuristics, but the <video> preview inserted fails to play and show due to a
style rule setting a "display: none" style if the media-query "pointer: coarse" evaluates to true.

In order to improve this website's behavior on iPadOS, we add a new quirk that prevents the "pointer: coarse" media query from evaluating to true.
This new quirk, shouldPreventPointerMediaQueryFromEvaluatingToCoarse(), evaluates to true only for this this website.

* css/MediaQueryEvaluator.cpp:
(WebCore::pointerEvaluate):
* page/Quirks.cpp:
(WebCore::Quirks::shouldPreventPointerMediaQueryFromEvaluatingToCoarse const):
* page/Quirks.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259387 268f45cc-cd09-0410-ab3c-d52691b4dbfc