Add some more scroll-latching tests https://bugs.webkit.org/show_bug.cgi?id=209924
Reviewed by Zalan Bujtas.

Add three scroll latching tests.

iframe-latch-small-deltas.html tests that starting a second scroll where the dominant
direction can't be determined re-uses the previous latching. This test currently fails.

overflow-in-iframe-latching.html tests a wheel over a scrolled-to-top overflow inside
a not-scrolled-to-top iframe; the iframe should scroll. This test currently fails.

latching-and-wheel-events.html tests that wheel events always go to the innermost target,
even if not latched, and don't propagate across frame boundaries. This test passes
the event propagation part, but fails in a similar way to overflow-in-iframe-latching.html .

* fast/scrolling/latching/iframe-latch-small-deltas-expected.txt: Added.
* fast/scrolling/latching/iframe-latch-small-deltas.html: Added.
* fast/scrolling/latching/latching-and-wheel-events-expected.txt: Added.
* fast/scrolling/latching/latching-and-wheel-events.html: Added.
* fast/scrolling/latching/overflow-in-iframe-latching-expected.txt: Added.
* fast/scrolling/latching/overflow-in-iframe-latching.html: Added.
* platform/mac-wk1/fast/scrolling/latching/iframe-latch-small-deltas-expected.txt: Added.
* platform/mac-wk1/fast/scrolling/latching/latching-and-wheel-events-expected.txt: Added.
* platform/mac-wk1/fast/scrolling/latching/overflow-in-iframe-latching-expected.txt: Added.
* resources/ui-helper.js:
(window.UIHelper.async mouseWheelScrollAt):
(window.UIHelper.async waitForScrollCompletion):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259417 268f45cc-cd09-0410-ab3c-d52691b4dbfc