Handle errors when grabbing grabbing microphone audio samples from the AudioUnit https://bugs.webkit.org/show_bug.cgi?id=210185
Reviewed by Eric Carlson.

We compute the buffer size to copy microphone samples when setting up the Audio Unit.
This is based on the preferred buffer size and sample rate.
But these values might change over time by the web page durig the capture.
If the preferred buffer size increases (for instance if the page stops using WebAudio), our buffer might be too small.
Capture will fail but we will not notify the web application.

Update the code to reconfigure the AudioUnit if AudioUnitRender returns an error of type kAudio_ParamError.
Update the code to only increment the number of microphoneProcsCalled if AudioUnitRender succeeds.
This will ensure that, should AudioUnitRender fails for some time, the timer will kick in and fail the capture.
Page will be notified and can call getUserMedia again to restart capture.

* platform/mediastream/mac/CoreAudioCaptureSource.cpp:
(WebCore::CoreAudioSharedUnit::processMicrophoneSamples):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259716 268f45cc-cd09-0410-ab3c-d52691b4dbfc