Fix handling non-linearized PDFs when incremental PDF loading is enabled. <rdar://problem/60619506> and https://bugs.webkit.org/show_bug.cgi?id=210208
Reviewed by Tim Horton (and I think Geoff Garen, probably. It's confusing.)

When we try to load a non-linearized PDF with PDFKit, it makes an outlandishly large range request
to try to verify the PDF file size.

That's covered by <rdar://problem/61473378>

Meanwhile we need to detect that and fallback to non-incremental PDF loading.

* WebProcess/Plugins/PDF/PDFPlugin.h:
* WebProcess/Plugins/PDF/PDFPlugin.mm:
(WebKit::PDFPlugin::receivedInvalidRangeRequest):
(WebKit::dataProviderGetBytesAtPositionCallback):
(WebKit::PDFPlugin::threadEntry):
(WebKit::PDFPlugin::installPDFDocument):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259760 268f45cc-cd09-0410-ab3c-d52691b4dbfc