REGRESSION (r258850) http/tests/frame-throttling/timer-throttle-in-cross-origin-subframe.html is flaky https://bugs.webkit.org/show_bug.cgi?id=210358 <rdar://problem/61266282>
Reviewed by Darin Adler.

r258850 ported window.postMessage() to the HTML event loop. Because event loops are per security origin,
the ordering of messages from the same-origin & cross-origin iframes is no longer guaranteed. This led
to flakiness due to message logging in the output but the test did not actually require a particular
ordering otherwise. To address the flakiness, this patch silences the logging of messages received.

* http/tests/frame-throttling/timer-throttle-in-cross-origin-subframe-expected.txt:
* http/tests/frame-throttling/timer-throttle-in-cross-origin-subframe.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259907 268f45cc-cd09-0410-ab3c-d52691b4dbfc