Enable the ability to build the ASM LLInt for ARMv7k. https://bugs.webkit.org/show_bug.cgi?id=210412
Reviewed by Sam Weinig.

Source/JavaScriptCore:

Fix the offlineasm so that it can build the ASM LLInt for ARMv7k.  This patch does
not actually enable the ASM LLInt.  The ARMv7k port still build the C Loop LLInt.

Also, the ARMv7k ASM LLInt is still broken and needs additional work before it
can run.  This patch only fixes things so that it will build.

* JavaScriptCore.xcodeproj/project.pbxproj:
- Added generate_settings_extractor.rb to the project so that we can view it from
  inside Xcode.

* offlineasm/arm.rb:
- Added support for the globaladdr LLInt instruction for ARMv7k.

* offlineasm/backends.rb:
- Fix the backend to enable ARMV7 also when building for ARMv7k.

Source/WTF:

Remove some old code that forces the ENABLE(C_LOOP) to be true if ENABLE(JIT) is
false.  These 2 options used to be mutually exclusive, but is no more.  Now, we
allow platforms to choose to build the ASM LLInt instead even if ENABLE(JIT) is
false.

* wtf/PlatformEnable.h:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259996 268f45cc-cd09-0410-ab3c-d52691b4dbfc